+++ src/App.vue
@@ -7,12 +7,14 @@
         <router-view />
       </transition>
     </v-content>
+    <Footer />
   </v-app>
 </template>
 
 <script>
 import Toolbar from '@/components/core/Toolbar.vue'
 import Loading from '@/components/core/Loading.vue'
+import Footer from '@/components/core/Footer.vue'
 
 export default {
   name: 'App',
@@ -57,7 +59,8 @@ export default {
   },
   components: {
     Toolbar,
-    Loading
+    Loading,
+    Footer
   },
   computed: {
     appTitle() {