+++ src/components/About.vue
@@ -4,6 +4,7 @@
       <Heading :title="$t('about.TITLE')" />
       <Description :description="$t('about.DESCRIPTION')" />
     </v-layout>
+    <ProjectDescription />
   </v-container>
 </template>